pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mccastellazzob"
version = "1.0.0"
description = "Sito dinamico motoclub MC Castellazzo - CodeRedCMS/Wagtail"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "MC Castellazzo", email = "info@mccastellazzob.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django :: 5.2",
    "Framework :: Wagtail :: 7",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "Django>=5.2,<5.3",
    "wagtail>=7.0,<8.0",
    "coderedcms>=6.0,<7.0",
    "wagtail-localize>=1.9,<2.0",
    "django-allauth>=65.0,<66.0",
    "psycopg[binary]>=3.1,<4.0",
    "gunicorn>=21.0,<22.0",
    "whitenoise>=6.6,<7.0",
    "Pillow>=10.0,<11.0",
    "django-jinja>=2.11,<3.0",
    "requests>=2.31,<3.0",
    "python-dotenv>=1.0,<2.0",
    "deep-translator>=1.11,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-django>=4.8,<5.0",
    "pytest-cov>=4.1,<5.0",
    "factory-boy>=3.3,<4.0",
    "faker>=22.0,<23.0",
    "ruff>=0.1,<1.0",
    "pre-commit>=3.6,<4.0",
    "django-debug-toolbar>=4.2,<5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mccastellazzob*", "apps*"]

[tool.ruff]
target-version = "py311"
line-length = 120
exclude = [
    ".git",
    "__pycache__",
    "migrations",
    "venv",
    ".venv",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["mccastellazzob", "apps"]
known-third-party = ["django", "wagtail", "coderedcms"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mccastellazzob.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
